Dashboard desktop and mobile responsive
