+++ src/shared/Navbar.jsx
@@ -1,4 +1,4 @@
-import React, { useState } from "react";
+import  { useState } from "react";
 import { Link } from "react-router-dom";
 import { GiHamburgerMenu } from "react-icons/gi"; 
 import { FaArrowAltCircleUp } from "react-icons/fa";
